worker/requirements.txt
albumentations
numpy
scipy
azure-storage
Flask
Flask-Cors
Pillow
python-dotenv
redis
requests
